\name{analogy}
\alias{analogy}
\title{Calculate the analogy relationship.}
\usage{
analogy (unitm,word1,word2,word3,num=20)
}
\arguments{
\item{unitm}{The unit word matrix.}

\item{word1}{The first word.}
\item{word2}{The second words.}
\item{word3}{The third words.}
\item{num}{The number of returned items.}
}

\value{
	The analogy words list for choosen.
}
\description{
Calculate the analogy relationship.
}
\author{
Whisly Wei <\email{whisly.wei@sap.com}>
}

