% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bin_to_txt}
\alias{bin_to_txt}
\title{to convert binary file to text file}
\usage{
bin_to_txt(bin_file, txt_file)
}
\arguments{
\item{bin_file}{binary file name}

\item{txt_file}{text file name}
}
\value{
text file giving text file
}
\description{
to convert binary file to text file
}
\examples{
bin_to_txt("binary_name.bin","txt_name.bin")
}

