\name{cosdist}
\alias{cosdist}
\title{Calculate the cosine distance of the word.}
\usage{
cosdist(unitm,word,num=20)
}
\arguments{
\item{unitm}{The unit matrix.}

\item{word}{The search word.}

\item{num}{The max return words.}
}
\value{
	Return an vector, sorted by distance.
}
\description{
	Calculate the cosine distance of the word.
}
\author{
Whisly Wei <\email{stweiyy@163.com}>
}

