% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance R API}
\usage{
distance(fileName = "", word = "", n = 100)
}
\arguments{
\item{fileName}{}

\item{word}{}

\item{n}{}
}
\description{
TODO
}
\author{
HatMatrix42
}

