\name{loadvector}
\alias{loadvector}
\title{load the vector of the model}
\usage{
	loadvector(modelfile)
}
\arguments{
\item{modelfile}{The file name of the model.}
}
\value{
The word vector matrix, the colname is the word, each row is a vector.
}
\description{
	This function load the vector of the word from the model file. The result is the word matrix. but pay attension that this is the original matrix ,not unit matrix.
}
\author{
whisly <stweiyy@163.com>
}
