\name{unitmatrix}
\alias{unitmatrix}
\title{Change the matrix to unit matrix.}
\usage{
unitmatrix(m)
}
\arguments{
\item{m}{The loaded matrix.}
}
\value{
The word matrix.
}
\description{
	Change the loaded matrix into an unit matrix.
}
\author{
Whisly Wei <\email{stweiyy@163.com}>
}
\references{
\url{https://code.google.com/p/word2vec/}
}

