% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove.R
\name{vocab_count}
\alias{vocab_count}
\title{Gets word count from the text data}
\usage{
vocab_count(file_name, vocab_file, verbose = 2, max_vocab = 0,
  min_count = 1)
}
\arguments{
\item{file_name}{text data}

\item{vocab_file}{output vocab count file name}

\item{max_vocab}{maximum vocab}

\item{min_count}{minimum occurrence of word}
}
\value{
data frame giving word count
}
\description{
Gets word count from the text data
}
\examples{
data=vocab_count("text8","vocab.txt")
}

