% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word2phrase}
\alias{word2phrase}
\title{to convert words to phrases}
\usage{
word2phrase(train_file, output_file, debug_mode = 0, min_count = 0,
  threshold = 0)
}
\arguments{
\item{train_file}{train file name}

\item{output_file}{output file name}

\item{debug_mode}{debug mode}

\item{min_count}{minimum count}

\item{threshold}{threshold value}
}
\value{
text file with unigrams and bigrams
}
\description{
to convert words to phrases
}
\examples{
\dontrun{
model=word2phrase("text8","vec.txt")
}
}

