% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word2vec}
\alias{word2vec}
\title{to train word2vec model}
\usage{
word2vec(layer1_size = 0, train_file = 0, save_vocab_file = 0,
  read_vocab_file = 0, debug_mode = 0, binary = 1, cbow = 0,
  alpha = 0, output_file = 0, window = 0, sample = 0, hs = 0,
  negative = 0, iter = 0, min_count = 0, classes = 0)
}
\arguments{
\item{layer1_size}{feature size}

\item{train_file}{training text data}

\item{save_vocab_file}{file name to save vocab data}

\item{read_vocab_file}{file name to read vocab}

\item{debug_mode}{debug mode}

\item{binary}{output file format}

\item{cbow}{training method used}

\item{alpha}{alpha}

\item{output_file}{output file name (MUST be binary)}

\item{window}{window}

\item{sample}{sample size}

\item{hs}{hs}

\item{negative}{negative}

\item{iter}{iter}

\item{min_count}{minimum count}

\item{classes}{classes}
}
\description{
to train text data
}
\examples{
\dontrun{
model=word2vec(train_file = "text8",output_file = "vec.bin",binary=1)
}
}

