% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word2vex}
\alias{word2vex}
\title{word2vex R API}
\usage{
word2vex(train_file = "", output_file = "", save_vocab_file = "",
  read_vocab_file = "", layer1_size = 100, binary = FALSE, cbow = TRUE,
  alpha = ifelse(cbow, 0.025, 0.05), window = 5, hs = 0, sample = 10^-3,
  negative = 5, num_threads = 4, iter = 5, min_count = 5, classes = 0)
}
\arguments{
\item{train_file}{}

\item{output_file}{}

\item{save_vocab_file}{}

\item{layer1_size}{Set size of word vectors; default is 100}

\item{binary}{Save the resulting vectors in binary moded,default is FALSE}

\item{cbow}{Use the continuous bag of words model,default is TRUE,
Using skip-gram model is FALSE}

\item{alpha}{Set the starting learning rate; default is 0.025 for
skip-gram and 0.05 for CBOW}

\item{window}{Set max skip length between words,default is 5}

\item{hs}{tUse Hierarchical Softmax,default is 0(not used)}

\item{sample}{Set threshold for occurrence of words.
Those that appear with higher frequency in the training data.}

\item{negative}{Number of negative examples; default is 5,
common values are 3 - 10 (0 = not used)}

\item{num_threads}{Use num_threads threads (default 4)}

\item{iter}{Run more training iterations (default 5)}

\item{min_count}{This will discard words that appear less than min_count times;}

\item{classes}{Output word classes rather than word vectors;
default number of classes is 0 (vectors are written)}
}
\description{
word2vec
}
\author{
HatMatrix42
}

