% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word_analogy}
\alias{word_analogy}
\title{Gets closest analogy to the searched word}
\usage{
word_analogy(file_name, search_words, num = 20)
}
\arguments{
\item{file_name}{trained data file name (MUST be binary)}

\item{num}{number of closest words}

\item{search_word}{search keywords}
}
\value{
data frame giving closest analogy and cosine distances
}
\description{
Gets closest analogy to the searched word
}
\examples{
ana=word_analogy("vec.bin","king queen man",20)
}

