% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{distance}
\alias{distance}
\title{Gets closest words to the searched word}
\usage{
distance(file_name, search_word, num = 15)
}
\arguments{
\item{file_name}{trained data file name (MUST be binary)}

\item{search_word}{search keyword}

\item{num}{number of closest words}
}
\value{
data frame giving closest words and cosine distances
}
\description{
Gets closest words to the searched word
}
\examples{
\dontrun{
dist=distance("vec.bin","king",20)
}
}

